refactor(insights): extract shared chrome components and SCSS for cross-tab reuse
Per the pre-Tab-7 audit. Two extractions, both mechanical, no
behavior change.

1. MetricCard + format helpers moved from tabs/subscribers/ to a
   new tabs/components/ directory. Both are generic — MetricCard's
   props were already free of any subscriber-specific shape and the
   format helpers (currency / number / percent / delta / tone) are
   pure pass-through utilities every future tab will need.

   Updated import paths:
     ScorecardSection  -> '../components/MetricCard'
     WindowedSection   -> '../components/MetricCard'
     PerformanceSection -> '../components/format'
     MetricCard's internal './format' import stays relative (same
       new directory).

2. tabs/subscribers/subscribers.scss was ~80% generic Insights chrome
   despite living in a tab-scoped file. Split into:

     tabs/components/sections.scss
       - __tab-loading, __tab-error, __tab-error-detail
       - __section, __section-heading, __section-caption,
         __section-empty
       - __metric-grid
       - __metric-card and all its sub-rules (-label, -body,
         -value, -delta with tones, -description)
       - __table-wrap and __table (incl. -num, the
         __table-row--variation + td:first-child indent pattern)

     tabs/subscribers/subscribers.scss (now slim, Tab 6 only)
       - __subscribers-tab orchestrator wrapper
       - __tenure-card (the percentile callouts container)
       - __stats-summary (the dl layout inside the tenure card)
       - __tenure-narrative

   style.scss now @use's the shared sections.scss so the chrome
   ships once with the main wizard bundle (insights-wizard.css)
   instead of inside the lazy-loaded subscribers chunk. Verified:
   metric-card / section / table / variation-row rules land in
   insights-wizard.css; tenure-card lands in the subscribers chunk
   (3352.css).

Tab 7 will inherit the chrome without importing it and only ship
its own tab-specific styles in its own lazy chunk.

Verified: build green, lint clean, REST endpoint unchanged
(active=35, mrr=738, performance rows=4 with the same nested
variation structure).
