+++ plugins/newspack-plugin/src/wizards/insights/tabs/components/format.ts
