+++ plugins/newspack-plugin/src/wizards/insights/tabs/components/sections.scss
@@ -1,0 +1,250 @@
+/**
+ * Newspack Insights — shared tab chrome (NPPD-1602/1616)
+ *
+ * Generic visuals that every Insights tab inherits: section wrappers
+ * and headings, the metric card chrome + responsive grid, the table
+ * treatment (incl. nested-variation rows), and tab-level loading /
+ * error blocks. Tab-specific styles (e.g. Tab 6's tenure card,
+ * Tab 7's donor-cohort layout) live alongside their tab orchestrator.
+ *
+ * Loaded by `src/wizards/insights/style.scss` so all tabs pick up
+ * the chrome without each having to import it.
+ */
+
+@use "~@wordpress/base-styles/colors" as wp-colors;
+
+.newspack-insights {
+
+	// Tab-level status blocks shown by orchestrators between data
+	// fetch and first render.
+
+	&__tab-loading,
+	&__tab-error {
+		padding: 32px 24px;
+		text-align: center;
+		font-size: 14px;
+		color: wp-colors.$gray-700;
+	}
+
+	&__tab-error {
+		color: wp-colors.$alert-red;
+
+		&-detail {
+			margin-top: 8px;
+			font-size: 13px;
+			color: wp-colors.$gray-600;
+			font-family: monospace;
+		}
+	}
+
+	// Sections are pure spacing — no background, no border. The cards
+	// inside them carry the chrome (per spec).
+
+	&__section {
+		display: flex;
+		flex-direction: column;
+		gap: 16px;
+	}
+
+	&__section-heading {
+		font-size: 18px;
+		font-weight: 600;
+		line-height: 1.3;
+		margin: 0;
+		color: wp-colors.$gray-900;
+	}
+
+	&__section-caption {
+		margin: 0;
+		font-size: 13px;
+		font-weight: 400;
+		color: wp-colors.$gray-700;
+	}
+
+	&__section-empty {
+		margin: 0;
+		padding: 20px;
+		background-color: #fff;
+		border: 1px solid wp-colors.$gray-200;
+		border-radius: 4px;
+		font-size: 14px;
+		font-weight: 400;
+		color: wp-colors.$gray-700;
+	}
+
+	// Scorecard grid — responsive: auto-fill with a 220px minimum per
+	// column. Each card stretches to fill its column.
+
+	&__metric-grid {
+		display: grid;
+		grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
+		gap: 16px;
+	}
+
+	// Card chrome per spec: white background, gray-200 border, 4px
+	// radius. Padding 24px 28px. Min-height 180px so cards line up
+	// uniformly across the grid AND give the centered hero value room
+	// to breathe between label and description.
+	//
+	// Internal layout (three regions, fixed order):
+	//   label  — top, left-aligned (spec's uppercase chip)
+	//   body   — flex: 1, hero sits at the top of this region,
+	//            left-aligned. Vertical position consistent across
+	//            cards because all cards share the same label height
+	//            + body padding.
+	//   description — pinned to the bottom edge, left-aligned
+	//
+	// Every card carries the brand-color top accent so a row reads as
+	// a single coherent unit.
+
+	&__metric-card {
+		display: flex;
+		flex-direction: column;
+		padding: 24px 28px;
+		background-color: #fff;
+		border: 1px solid wp-colors.$gray-200;
+		border-top: 3px solid var(--wp-admin-theme-color);
+		border-radius: 4px;
+		min-height: 180px;
+
+		&-label {
+			font-size: 12px;
+			font-weight: 600;
+			line-height: 1.4;
+			letter-spacing: 0.05em;
+			text-transform: uppercase;
+			color: wp-colors.$gray-700;
+			// Reserve two lines of vertical space so a long label that
+			// wraps ("MONTHLY RECURRING REVENUE") doesn't push its hero
+			// number down relative to neighboring cards whose labels fit
+			// on one line. 12px font × 1.4 line-height × 2 lines.
+			min-height: calc(12px * 1.4 * 2);
+		}
+
+		&-body {
+			flex: 1 1 auto;
+			display: flex;
+			flex-direction: column;
+			justify-content: flex-start; // hero sits at the top, just under the label
+			align-items: flex-start;     // left-aligned to match label
+			gap: 8px;
+			padding: 16px 0 0;
+		}
+
+		&-value {
+			margin: 0;
+			// Explicit font stack + lockdown to prevent the wp-admin chrome's
+			// inherited font from flipping between cards (the bug was: digits
+			// in "35" rendered slim, "$738.33" rendered heavier with a slight
+			// optical slant — different chars resolving in different
+			// fallback fonts because no explicit family was declared). The
+			// stack matches @wordpress/base-styles' admin body font.
+			font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+				Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
+			font-size: 44px;
+			font-weight: 500;
+			font-style: normal;
+			line-height: 1.05;
+			letter-spacing: -0.01em;
+			color: wp-colors.$gray-900;
+			font-variant-numeric: tabular-nums;
+			font-feature-settings: "tnum" 1;
+			font-synthesis: none;
+			text-align: left;
+		}
+
+		&-delta {
+			font-size: 13px;
+			font-weight: 600;
+			font-variant-numeric: tabular-nums;
+
+			// Spec calls for newspack-colors $success-600 / $error-600.
+			// Chrome stylesheet only imports @wordpress/base-styles, so
+			// using its alert tokens here keeps the import surface tight;
+			// swap to newspack-colors when (if) those tokens get wired in.
+			&--positive {
+				color: wp-colors.$alert-green;
+			}
+
+			&--negative {
+				color: wp-colors.$alert-red;
+			}
+
+			&--neutral {
+				color: wp-colors.$gray-600;
+			}
+		}
+
+		&-description {
+			margin: 16px 0 0;
+			font-size: 13px;
+			font-weight: 400;
+			line-height: 1.4;
+			color: wp-colors.$gray-700;
+			text-align: left;
+		}
+	}
+
+	// Table — card wraps the table, table fills (padding 0 on
+	// wrapper). Header bg per spec gray-100. The variation row pattern
+	// is generic: any nested parent/child data shape can use it.
+
+	&__table-wrap {
+		overflow-x: auto;
+		background-color: #fff;
+		border: 1px solid wp-colors.$gray-200;
+		border-radius: 4px;
+	}
+
+	&__table {
+		width: 100%;
+		border-collapse: collapse;
+		font-size: 14px;
+
+		th,
+		td {
+			padding: 12px 20px;
+			text-align: left;
+			border-bottom: 1px solid wp-colors.$gray-200;
+		}
+
+		th {
+			font-size: 12px;
+			font-weight: 600;
+			letter-spacing: 0.04em;
+			text-transform: uppercase;
+			color: wp-colors.$gray-700;
+			background-color: wp-colors.$gray-100;
+			border-bottom: 1px solid wp-colors.$gray-300;
+		}
+
+		tbody tr:last-child td {
+			border-bottom: 0;
+		}
+
+		&-num {
+			text-align: right;
+			font-variant-numeric: tabular-nums;
+		}
+
+		// Nested variation/child rows sit directly under their parent.
+		// Same chrome as parent rows, no background change, just a
+		// slightly lighter text color (gray-700) and an indent on the
+		// first cell so the nesting reads at a glance.
+		//
+		// The indent rule targets `td:first-child` rather than a
+		// per-cell modifier class because the modifier would be
+		// specificity (0,1,0) and lose to the base `.table td` rule
+		// (0,2,1). `&-row--variation td:first-child` is (0,2,2) which
+		// wins the cascade cleanly.
+		&-row--variation {
+			td {
+				color: wp-colors.$gray-700;
+			}
+
+			td:first-child {
+				padding-left: 44px; // 20px table padding + 24px indent
+			}
+		}
+	}
+}
